add map to survey study area, refactoring
