+++ app/src/features/projects/components/ProjectLocationForm.tsx
@@ -1,36 +1,17 @@
 import Box from '@material-ui/core/Box';
-import Button from '@material-ui/core/Button';
-import Tooltip from '@material-ui/core/Tooltip';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
-import Typography from '@material-ui/core/Typography';
-import makeStyles from '@material-ui/core/styles/makeStyles';
-//@ts-ignore
-import { kml } from '@tmcw/togeojson';
-import shp from 'shpjs';
-import bbox from '@turf/bbox';
 import {
   default as MultiAutocompleteFieldVariableSize,
   IMultiAutocompleteFieldOption
 } from 'components/fields/MultiAutocompleteFieldVariableSize';
-import MapContainer from 'components/map/MapContainer';
 import { useFormikContext } from 'formik';
 import { Feature } from 'geojson';
 import React, { useEffect, useState } from 'react';
 import yup from 'utils/YupSchema';
-import { v4 as uuidv4 } from 'uuid';
 import Link from '@material-ui/core/Link';
-
-const useStyles = makeStyles({
-  bold: {
-    fontWeight: 'bold'
-  },
-  uploadButton: {
-    border: '2px solid',
-    textTransform: 'capitalize',
-    fontWeight: 'bold'
-  }
-});
+import MapBoundary from 'components/boundary/MapBoundary';
+import { updateMapBounds } from 'utils/mapBoundaryUploadHelpers';
 
 export interface IProjectLocationForm {
   regions: string[];
@@ -59,8 +40,6 @@ export interface IProjectLocationFormProps {
  * @return {*}
  */
 const ProjectLocationForm: React.FC<IProjectLocationFormProps> = (props) => {
-  const classes = useStyles();
-
   const formikProps = useFormikContext<IProjectLocationForm>();
 
   const { values, touched, errors, handleChange, handleSubmit, setFieldValue } = formikProps;
@@ -71,91 +50,8 @@ const ProjectLocationForm: React.FC<IProjectLocationFormProps> = (props) => {
 
   useEffect(() => {
     setIsLoading(false);
-
-    /*
-      If no geometries, we do not need to set bounds
-
-      If there is only one geometry and it is a point, we cannot do the bound setting
-      because leaflet does not know how to handle that and tries to zoom in way too much
-
-      If there are multiple points or a polygon and a point, this is not an issue
-    */
-    if (!values.geometry.length || (values.geometry.length === 1 && values.geometry[0].geometry.type === 'Point')) {
-      return;
-    }
-
-    const allGeosFeatureCollection = {
-      type: 'FeatureCollection',
-      features: [...values.geometry]
-    };
-    const bboxCoords = bbox(allGeosFeatureCollection);
-
-    setBounds([
-      [bboxCoords[1], bboxCoords[0]],
-      [bboxCoords[3], bboxCoords[2]]
-    ]);
+    updateMapBounds(values, setBounds);
   }, [values.geometry]);
-
-  /**
-   * Convert a zipped shapefile to geojson
-   * @param e The file upload event
-   */
-  const handleShapefileUpload = (e: any) => {
-    // Only accept one file
-    const file = e.target.files[0];
-
-    // Back out if not a zipped file
-    if (!file?.type.match(/zip/)) {
-      return;
-    }
-
-    // Create a file reader to extract the binary data
-    const reader = new FileReader();
-    reader.readAsArrayBuffer(file);
-
-    // When the file is loaded run the conversion
-    reader.onload = async (event: any) => {
-      // The converter wants a buffer
-      const zip: Buffer = event?.target?.result as Buffer;
-
-      // Exit out if no zip
-      if (!zip) {
-        return;
-      }
-
-      // Run the conversion
-      const geojson = await shp(zip);
-      const features = (geojson as any).features;
-      setFieldValue('geometry', [...features, ...values.geometry]);
-    };
-  };
-
-  const handleKMLUpload = async (e: any) => {
-    setIsLoading(true);
-
-    const file = e.target.files[0];
-    const fileAsString = await file?.text().then((xmlString: string) => {
-      return xmlString;
-    });
-
-    if (file?.type !== 'application/vnd.google-earth.kml+xml' && !fileAsString?.includes('</kml>')) {
-      setUploadError('You must upload a KML file, please try again.');
-      setIsLoading(false);
-      return;
-    }
-
-    const domKml = new DOMParser().parseFromString(fileAsString, 'application/xml');
-    const geojson = kml(domKml);
-
-    let sanitizedGeoJSON: Feature[] = [];
-    geojson.features.forEach((feature: Feature) => {
-      if (feature.geometry) {
-        sanitizedGeoJSON.push(feature);
-      }
-    });
-
-    setFieldValue('geometry', [...sanitizedGeoJSON, ...values.geometry]);
-  };
 
   return (
     <form onSubmit={handleSubmit}>
@@ -190,61 +86,17 @@ const ProjectLocationForm: React.FC<IProjectLocationFormProps> = (props) => {
             helperText={touched.location_description && errors.location_description}
           />
         </Grid>
-        <Grid item xs={12}>
-          <Typography className={classes.bold}>Project Boundary</Typography>
-          <Box display="flex" mt={3}>
-            <Tooltip arrow color="secondary" title="Will only accept kml files, kmz files not accepted.">
-              <Button
-                variant="outlined"
-                component="label"
-                size="medium"
-                color="primary"
-                disabled={isLoading}
-                onClick={() => setUploadError('')}
-                className={classes.uploadButton}>
-                <input
-                  key={uuidv4()}
-                  data-testid="file-upload"
-                  type="file"
-                  hidden
-                  onChange={(e) => handleKMLUpload(e)}
-                />
-                Upload KML
-              </Button>
-            </Tooltip>
-            <Tooltip arrow color="secondary" title="Will only accept zipped shapefiles of a known projection.">
-              <Button
-                variant="outlined"
-                component="label"
-                size="medium"
-                color="primary"
-                disabled={isLoading}
-                className={classes.uploadButton}
-                style={{ marginLeft: '1rem' }}>
-                <input
-                  key={uuidv4()}
-                  data-testid="shp-upload"
-                  type="file"
-                  hidden
-                  onChange={(e) => handleShapefileUpload(e)}
-                />
-                Upload Shapefile
-              </Button>
-            </Tooltip>
-          </Box>
-          <Box mt={2}>{uploadError && <Typography style={{ color: '#db3131' }}>{uploadError}</Typography>}</Box>
-          <Box mt={5} height={500}>
-            <MapContainer
-              mapId="project_location_form_map"
-              //@ts-ignore
-              geometryState={{
-                geometry: values.geometry,
-                setGeometry: (newGeo: Feature[]) => setFieldValue('geometry', newGeo)
-              }}
-              bounds={bounds}
-            />
-          </Box>
-        </Grid>
+        <MapBoundary
+          title="Project Boundary"
+          mapId="project_location_form_map"
+          isLoading={isLoading}
+          setIsLoading={setIsLoading}
+          uploadError={uploadError}
+          setUploadError={setUploadError}
+          values={values}
+          bounds={bounds}
+          setFieldValue={setFieldValue}
+        />
       </Grid>
     </form>
   );
