+++ app/src/features/surveys/components/StudyAreaForm.tsx
@@ -2,19 +2,24 @@ import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
 import AutocompleteField, { IAutocompleteFieldOption } from 'components/fields/AutocompleteField';
 import { useFormikContext } from 'formik';
-import React from 'react';
+import { Feature } from 'geojson';
+import React, { useEffect, useState } from 'react';
 import yup from 'utils/YupSchema';
+import MapBoundary from 'components/boundary/MapBoundary';
+import { updateMapBounds } from 'utils/mapBoundaryUploadHelpers';
 
 export interface IStudyAreaForm {
   survey_area_name: string;
   park: string;
   management_unit: string;
+  geometry: Feature[];
 }
 
 export const StudyAreaInitialValues: IStudyAreaForm = {
   survey_area_name: '',
   park: '',
-  management_unit: ''
+  management_unit: '',
+  geometry: []
 };
 
 export const StudyAreaYupSchema = yup.object().shape({
@@ -34,7 +39,16 @@ export interface IStudyAreaFormProps {
  * @return {*}
  */
 const StudyAreaForm: React.FC<IStudyAreaFormProps> = (props) => {
-  const { values, touched, errors, handleChange } = useFormikContext<IStudyAreaForm>();
+  const { values, touched, errors, handleChange, setFieldValue } = useFormikContext<IStudyAreaForm>();
+
+  const [bounds, setBounds] = useState<any>([]);
+  const [uploadError, setUploadError] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
+
+  useEffect(() => {
+    setIsLoading(false);
+    updateMapBounds(values, setBounds);
+  }, [values.geometry]);
 
   return (
     <form>
@@ -65,6 +79,17 @@ const StudyAreaForm: React.FC<IStudyAreaFormProps> = (props) => {
             required={true}
           />
         </Grid>
+        <MapBoundary
+          title="Study Area Boundary"
+          mapId="study_area_form_map"
+          isLoading={isLoading}
+          setIsLoading={setIsLoading}
+          uploadError={uploadError}
+          setUploadError={setUploadError}
+          values={values}
+          bounds={bounds}
+          setFieldValue={setFieldValue}
+        />
       </Grid>
     </form>
   );
